warnings flag

+++ protocol/include/fractal.c
@@ -331,7 +331,7 @@ void* TryReadingTCPPacket( struct SocketContext* context )
 		// If the previous recvp was maxed out, then try pulling some more from recvp
 	} while( len == TCP_SEGMENT_SIZE );
 
-	if( reading_packet_len > sizeof( int ) )
+	if( (unsigned long) reading_packet_len > sizeof( int ) )
 	{
 		// The amount of data bytes read (actual len), and the amount of bytes we're looking for (target len), respectively
 		int actual_len = reading_packet_len - sizeof( int );
@@ -345,7 +345,7 @@ void* TryReadingTCPPacket( struct SocketContext* context )
 
 			// Move the rest of the read bytes to the beginning of the buffer to continue
 			int start_next_bytes = sizeof(int) + target_len;
-			for( int i = start_next_bytes; i < sizeof(int) + actual_len; i++ )
+			for( unsigned long i = start_next_bytes; i < sizeof(int) + actual_len; i++ )
 			{
 				reading_packet_buffer[i - start_next_bytes] = reading_packet_buffer[i];
 			}
@@ -593,7 +593,7 @@ uint32_t Hash(void* buf, size_t len)
 	}
 
 	uint32_t hash = (pre_hash << 32) ^ pre_hash;
-	for (int i = 0; i < len; ++i)
+	for (size_t i = 0; i < len; ++i)
 	{
 		hash += key[i];
 		hash += (hash << 10);
